# Conflicts:
#	src/main/java/site/woulduduo/controller/UserController.java
#	src/main/java/site/woulduduo/repository/BoardLikeRepository.java
#	src/main/java/site/woulduduo/service/UserService.java

+++ src/main/java/site/woulduduo/dto/riot/MostChampInfo.java
@@ -13,7 +13,7 @@ public class MostChampInfo {
     private String champName;
     private int winCount;
     private int loseCount;
-    private double winRate;
+    private int winRate;
     private double kda;
 
 }
